# backend/app/api/v1/projects.py
"""Projects and their encryption recipient keys (envelope §4.1, §4.3).

The keypair is generated in the browser and the private half is downloaded by
the user; only the public key is ever sent here. There is deliberately no
endpoint that generates a keypair server-side — the server would hold the
private key at the moment of creation, and `project_e2e` would be a promise the
architecture could not keep.
"""

from typing import Annotated, Any

from fastapi import APIRouter, Depends, File, Form, HTTPException, Path, Query, UploadFile
from sqlalchemy.ext.asyncio import AsyncSession

from app.api.access import access
from app.api.deps import get_db
from app.api.schemas import MessageError
from app.modules.entities import service as dataset_service
from app.modules.entities.schemas import DatasetRefusedError, PublishDatasetResponse
from app.modules.forms.xlsform.datasets import (
    DEFAULT_VALUE_COLUMN,
    CsvUnreadable,
    read_companion_csv,
)
from app.modules.media import service as media_service
from app.modules.media.schemas import MediaPolicyResponse, MediaPolicyUpdate
from app.modules.projects import service
from app.modules.projects.schemas import (
    ProjectKeyCreate,
    ProjectKeyDetail,
    ProjectKeyError,
    ProjectKeyErrorResponse,
    ProjectKeyListResponse,
    ProjectListResponse,
)

router = APIRouter()

# 404, 409 and 422 from the key endpoints all carry the same reasoned body.
# 422 is shared with FastAPI's own request-validation failure, which has a
# different shape; only one of the two can be declared under one status, and
# the framework's is the one a client meets first (see app/api/schemas.py).
_KEY_ERRORS: dict[int | str, dict[str, Any]] = {
    404: {"model": ProjectKeyErrorResponse},
    409: {"model": ProjectKeyErrorResponse},
}


@router.get(
    "",
    response_model=ProjectListResponse,
    response_model_by_alias=True,
    dependencies=[
        Depends(
            access(
                permission=(
                    "form.edit",
                    "submission.view",
                    "form.publish",
                    "project.manage",
                    "sample.upload",
                    "team.manage",
                )
            )
        )
    ],
)
async def list_projects(
    session: Annotated[AsyncSession, Depends(get_db)],
    include_archived: Annotated[bool, Query(alias="includeArchived")] = False,
) -> ProjectListResponse:
    """Every project, its security mode, and how many recipient keys it has.

    The key count is the operational fact: a project in an encrypting mode with
    zero active keys cannot receive data at all, because its devices refuse to
    push rather than send answers in the clear.
    """
    async with session.begin():
        return await service.list_projects(session, include_archived=include_archived)


@router.get(
    "/{project_id}/keys",
    response_model=ProjectKeyListResponse,
    response_model_by_alias=True,
    responses={404: {"model": MessageError}},
    dependencies=[Depends(access(permission="project.manage"))],
)
async def list_project_keys(
    session: Annotated[AsyncSession, Depends(get_db)],
    project_id: Annotated[str, Path(min_length=1, max_length=64)],
    include_revoked: Annotated[bool, Query(alias="includeRevoked")] = False,
) -> ProjectKeyListResponse:
    """The project's recipient keys — public halves only.

    Nothing here is secret. Every byte is useless without a private key the
    server has never held, which is what makes the mode worth having.
    """
    async with session.begin():
        keys = await service.list_project_keys(
            session, project_id, include_revoked=include_revoked
        )
    if keys is None:
        raise HTTPException(status_code=404, detail="project not found")
    return keys


@router.post(
    "/{project_id}/keys",
    response_model=ProjectKeyDetail,
    response_model_by_alias=True,
    status_code=201,
    responses=_KEY_ERRORS,
    dependencies=[Depends(access(permission="project.manage"))],
)
async def add_project_key(
    request: ProjectKeyCreate,
    session: Annotated[AsyncSession, Depends(get_db)],
    project_id: Annotated[str, Path(min_length=1, max_length=64)],
) -> ProjectKeyDetail:
    """Register a public key as a recipient for this project's submissions.

    Accepts a raw 32-byte X25519 public key, a role and a label, and nothing
    else: the request model forbids unknown fields, so a client that sends a
    private key under any name gets a 422 naming the field rather than a 201 and
    a secret sitting in a request log.

    Refuses a small-order point, which would make every wrap openable by anyone
    while looking like redundancy, and refuses a public key the project already
    holds, because a second copy of one key is not a second recipient (§4.3).
    """
    async with session.begin():
        try:
            return await service.add_project_key(session, project_id, request)
        except service.KeyRegistrationError as error:
            raise HTTPException(
                status_code=error.status_code,
                # Built from the declared model, not a bare dict, so the body
                # on the wire and the body in the contract have one author.
                detail=ProjectKeyError(
                    reason=error.reason, message=error.message
                ).model_dump(),
            ) from error


@router.post(
    "/{project_id}/keys/{key_id}/revoke",
    response_model=ProjectKeyDetail,
    response_model_by_alias=True,
    responses=_KEY_ERRORS,
    dependencies=[Depends(access(permission="project.manage"))],
)
async def revoke_project_key(
    session: Annotated[AsyncSession, Depends(get_db)],
    project_id: Annotated[str, Path(min_length=1, max_length=64)],
    key_id: Annotated[str, Path(min_length=1, max_length=64)],
) -> ProjectKeyDetail:
    """Retire a recipient key (encryption envelope §8).

    A POST rather than a DELETE, because nothing is deleted. The row stays and
    the wraps that name it stay: submissions collected while this key was active
    are still encrypted to it and always will be, and the console needs the row
    to say whose private key opens them. What changes is the future — no new
    submission is wrapped to it, and devices stop being offered it.

    Refuses to retire the last active recipient of an encrypting project, which
    would stop collection in the field without telling anyone. Idempotent: a
    second revoke returns the first one's timestamp rather than moving it.
    """
    async with session.begin():
        try:
            return await service.revoke_project_key(session, project_id, key_id)
        except service.KeyRegistrationError as error:
            raise HTTPException(
                status_code=error.status_code,
                # Built from the declared model, not a bare dict, so the body
                # on the wire and the body in the contract have one author.
                detail=ProjectKeyError(
                    reason=error.reason, message=error.message
                ).model_dump(),
            ) from error


@router.get(
    "/{project_id}/media-policy",
    response_model=MediaPolicyResponse,
    response_model_by_alias=True,
    responses={404: {"model": MessageError}},
    dependencies=[Depends(access(permission="project.manage"))],
)
async def get_media_policy(
    session: Annotated[AsyncSession, Depends(get_db)],
    project_id: Annotated[str, Path(min_length=1, max_length=64)],
) -> MediaPolicyResponse:
    """This project's capture settings. Devices read them from
    `GET /devices/{deviceId}/media-policy`; this is the console's view of the
    same three numbers."""
    async with session.begin():
        policy = await media_service.project_media_policy(session, project_id)
    if policy is None:
        raise HTTPException(status_code=404, detail="project not found")
    return policy


@router.patch(
    "/{project_id}/media-policy",
    response_model=MediaPolicyResponse,
    response_model_by_alias=True,
    responses={404: {"model": MessageError}},
    dependencies=[Depends(access(permission="project.manage"))],
)
async def set_media_policy(
    session: Annotated[AsyncSession, Depends(get_db)],
    project_id: Annotated[str, Path(min_length=1, max_length=64)],
    update: MediaPolicyUpdate,
) -> MediaPolicyResponse:
    """Change the capture settings. Omitted fields are left alone.

    Takes effect on each device at its next sync, and is not retroactive — nor
    meant to be. A photograph already captured at 1600px is the evidence that
    exists; re-compressing history would be inventing a different one.

    Lowering `gpsMaxAccuracyM` tightens what devices will accept from then on.
    Points already collected under the old threshold keep the accuracy they were
    recorded with, which is why the reading is stored beside every point rather
    than being checked and discarded.
    """
    async with session.begin():
        policy = await media_service.update_media_policy(session, project_id, update)
    if policy is None:
        raise HTTPException(status_code=404, detail="project not found")
    return policy


@router.post(
    "/{project_id}/datasets",
    response_model=PublishDatasetResponse,
    response_model_by_alias=True,
    status_code=201,
    responses={400: {"model": MessageError}, 409: {"model": DatasetRefusedError}},
    description=(
        "Publish one immutable version of a dataset from a CSV (Form IR §3).\n\n"
        "Multipart, not JSON, and deliberately: the caller already has the file "
        "— it uploaded the same bytes to POST /forms/import — and a 38,000-row "
        "village list re-encoded as a JSON array is several megabytes of body "
        "for no gain. It also keeps one CSV parser on the server rather than "
        "one here and one in every client.\n\n"
        "**Idempotent by content.** Re-uploading an unchanged file returns the "
        "existing version with `created: false` rather than publishing a "
        "duplicate — which matters because a console re-sends the companion "
        "files every time somebody presses Publish, and a device holding the "
        "old version id would otherwise be told it is behind and re-fetch rows "
        "that did not change.\n\n"
        "A published version never changes. Re-publishing a *numbered* version "
        "with different content is refused with a 409: a form version pinned to "
        "it would have its choice list moved underneath answers already "
        "collected against it."
    ),
    dependencies=[Depends(access(permission="form.edit"))],
)
async def publish_dataset(
    session: Annotated[AsyncSession, Depends(get_db)],
    project_id: Annotated[str, Path(min_length=1, max_length=64)],
    file: Annotated[UploadFile, File(description="The reference data, as CSV")],
    dataset_key: Annotated[
        str,
        Form(
            alias="datasetKey",
            min_length=1,
            max_length=200,
            description="The Form IR key this list is published under — what "
            "`choices.dataset` names. The XLSForm importer derives it from the "
            "file name and reports what it chose.",
        ),
    ],
    key_column: Annotated[
        str,
        Form(
            alias="keyColumn",
            min_length=1,
            max_length=200,
            description="The column holding each row's identity — what a "
            "`select_one_from_file` stores as the answer. Defaults to `name`, "
            "which is what XLSForm requires of these files.",
        ),
    ] = DEFAULT_VALUE_COLUMN,
    name: Annotated[
        str | None,
        Form(description="Display name for the dataset. Defaults to its key."),
    ] = None,
) -> PublishDatasetResponse:
    """Turn one CSV into an immutable dataset version."""
    data = await file.read()
    try:
        parsed = read_companion_csv(file.filename or f"{dataset_key}.csv", data)
    except CsvUnreadable as failure:
        # Not a domain refusal: there is no dataset here to refuse. Same
        # distinction as WorkbookError against a diagnostic.
        raise HTTPException(status_code=400, detail=str(failure)) from failure

    async with session.begin():
        try:
            published = await dataset_service.publish_dataset_version(
                session,
                project_id=project_id,
                dataset_key=dataset_key,
                rows=[dict(row) for row in parsed.rows],
                key_column=key_column,
                name=name,
            )
        except dataset_service.DatasetRefused as refusal:
            raise HTTPException(status_code=409, detail=refusal.reasons) from refusal

    return PublishDatasetResponse(
        dataset_id=published.dataset_id,
        dataset_version_id=published.dataset_version_id,
        dataset_key=published.dataset_key,
        version=published.version,
        row_count=published.row_count,
        checksum=published.checksum,
        created=published.created,
        # The file's own findings travel with the dataset's: a padded row and a
        # confusable key are the same kind of fact to whoever uploaded it.
        warnings=[*parsed.warnings, *published.warnings],
        published_at=published.published_at,
    )