# specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Data Collection Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/devices/{device_id}/media-policy": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "Media Policy",
        "description": "The capture settings this device must apply, and the chunk size.\n\nFetched every sync and cached locally, for the same reason as the crypto\nconfig: a device may go two weeks without a server and has to keep capturing\nto the project's settings throughout.\n\nTwo of the three are compression, which trades evidentiary quality against\nbandwidth — only the study knows which side it is on. The third,\n`gpsMaxAccuracyM`, is a refusal threshold rather than a preference: a phone\nindoors will report a two-kilometre \"fix\" with exactly the authority of a\ngood one, and nothing downstream can tell them apart afterwards. The client\nrefuses a worse fix rather than storing it quietly.\n\n404 when the device is unknown or revoked — a device the server will not\naccept data from has no business learning a project's settings either.",
        "operationId": "media_policy_api_v1_devices__device_id__media_policy_get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaPolicyResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "description": "Every project, its security mode, and how many recipient keys it has.\n\nThe key count is the operational fact: a project in an encrypting mode with\nzero active keys cannot receive data at all, because its devices refuse to\npush rather than send answers in the clear.",
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "name": "includeArchived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Includearchived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/keys": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Project Keys",
        "description": "The project's recipient keys — public halves only.\n\nNothing here is secret. Every byte is useless without a private key the\nserver has never held, which is what makes the mode worth having.",
        "operationId": "list_project_keys_api_v1_projects__project_id__keys_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Project Id"
            }
          },
          {
            "name": "includeRevoked",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Includerevoked"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectKeyListResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Add Project Key",
        "description": "Register a public key as a recipient for this project's submissions.\n\nAccepts a raw 32-byte X25519 public key, a role and a label, and nothing\nelse: the request model forbids unknown fields, so a client that sends a\nprivate key under any name gets a 422 naming the field rather than a 201 and\na secret sitting in a request log.\n\nRefuses a small-order point, which would make every wrap openable by anyone\nwhile looking like redundancy, and refuses a public key the project already\nholds, because a second copy of one key is not a second recipient (§4.3).",
        "operationId": "add_project_key_api_v1_projects__project_id__keys_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectKeyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectKeyDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectKeyErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectKeyErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/keys/{key_id}/revoke": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Revoke Project Key",
        "description": "Retire a recipient key (encryption envelope §8).\n\nA POST rather than a DELETE, because nothing is deleted. The row stays and\nthe wraps that name it stay: submissions collected while this key was active\nare still encrypted to it and always will be, and the console needs the row\nto say whose private key opens them. What changes is the future — no new\nsubmission is wrapped to it, and devices stop being offered it.\n\nRefuses to retire the last active recipient of an encrypting project, which\nwould stop collection in the field without telling anyone. Idempotent: a\nsecond revoke returns the first one's timestamp rather than moving it.",
        "operationId": "revoke_project_key_api_v1_projects__project_id__keys__key_id__revoke_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Project Id"
            }
          },
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectKeyDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectKeyErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectKeyErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/media-policy": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Media Policy",
        "description": "This project's capture settings. Devices read them from\n`GET /devices/{deviceId}/media-policy`; this is the console's view of the\nsame three numbers.",
        "operationId": "get_media_policy_api_v1_projects__project_id__media_policy_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaPolicyResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Set Media Policy",
        "description": "Change the capture settings. Omitted fields are left alone.\n\nTakes effect on each device at its next sync, and is not retroactive — nor\nmeant to be. A photograph already captured at 1600px is the evidence that\nexists; re-compressing history would be inventing a different one.\n\nLowering `gpsMaxAccuracyM` tightens what devices will accept from then on.\nPoints already collected under the old threshold keep the accuracy they were\nrecorded with, which is why the reading is stored beside every point rather\nthan being checked and discarded.",
        "operationId": "set_media_policy_api_v1_projects__project_id__media_policy_patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaPolicyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaPolicyResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/datasets": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Publish Dataset",
        "description": "Publish one immutable version of a dataset from a CSV (Form IR §3).\n\nMultipart, not JSON, and deliberately: the caller already has the file — it uploaded the same bytes to POST /forms/import — and a 38,000-row village list re-encoded as a JSON array is several megabytes of body for no gain. It also keeps one CSV parser on the server rather than one here and one in every client.\n\n**Idempotent by content.** Re-uploading an unchanged file returns the existing version with `created: false` rather than publishing a duplicate — which matters because a console re-sends the companion files every time somebody presses Publish, and a device holding the old version id would otherwise be told it is behind and re-fetch rows that did not change.\n\nA published version never changes. Re-publishing a *numbered* version with different content is refused with a 409: a form version pinned to it would have its choice list moved underneath answers already collected against it.",
        "operationId": "publish_dataset_api_v1_projects__project_id__datasets_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_publish_dataset_api_v1_projects__project_id__datasets_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishDatasetResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetRefusedError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeyRole": {
        "type": "string",
        "enum": [
          "primary",
          "backup",
          "recovery"
        ]
      },
...
      "MediaPolicyResponse": {
        "properties": {
          "projectId": {
            "type": "string",
            "title": "Projectid"
          },
          "chunkSize": {
            "type": "integer",
            "title": "Chunksize"
          },
          "policy": {
            "$ref": "#/components/schemas/MediaPolicy"
          }
        },
        "type": "object",
        "required": [
          "projectId",
          "chunkSize",
          "policy"
        ],
        "title": "MediaPolicyResponse"
      },
...
      "MessageError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "MessageError",
        "description": "The body of a refusal that carries prose and nothing to branch on.\n\n`{\"detail\": \"submission not found\"}`. Used where there is exactly one way\nto fail and the status code already says which: a 404 on a read endpoint.\nAnything a client must branch on gets a `reason` field instead — see\n`DeviceRegisterError` and `ProjectKeyError` in `modules/projects/schemas.py`."
      },
...
      "ProjectKeyCreate": {
        "properties": {
          "publicKey": {
            "type": "string",
            "title": "Publickey"
          },
          "role": {
            "$ref": "#/components/schemas/KeyRole"
          },
          "label": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "publicKey",
          "role",
          "label"
        ],
        "title": "ProjectKeyCreate",
        "description": "A public key being registered as a recipient (envelope §4.1).\n\n`extra=\"forbid\"`, deliberately. The private key is generated in the browser\nand downloaded by the user; it must never reach the server. A client that\nsends one — under any field name — gets a 422 naming the field rather than\na 201 and a silently ignored secret sitting in a request log.\n\nAn X25519 private key is 32 bytes and so is a public key, so no server can\ntell one from the other by looking. That is exactly why the guarantee has to\nbe structural: the server never asks for a private key, accepts no field\nthat could carry one, and refuses anything shaped like a key container."
      },
...
      "ProjectKeyErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ProjectKeyError"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ProjectKeyErrorResponse",
        "description": "404, 409 and 422 from the project key endpoints."
      },
      "ProjectKeyListResponse": {
        "properties": {
          "projectId": {
            "type": "string",
            "title": "Projectid"
          },
          "securityMode": {
            "$ref": "#/components/schemas/SecurityMode"
          },
          "keys": {
            "items": {
              "$ref": "#/components/schemas/ProjectKeyDetail"
            },
            "type": "array",
            "title": "Keys"
          }
        },
        "type": "object",
        "required": [
          "projectId",
          "securityMode",
          "keys"
        ],
        "title": "ProjectKeyListResponse"
      },
...
      "ProjectListResponse": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            },
            "type": "array",
            "title": "Projects"
          }
        },
        "type": "object",
        "required": [
          "projects"
        ],
        "title": "ProjectListResponse"
      },
...
      "SecurityMode": {
        "type": "string",
        "enum": [
          "standard",
          "field_level",
          "project_e2e"
        ]
      },
...
    }
  }
}